Delete the Calendar, MealEntry, Settings, and WeeklyCalendar storyboards
I deleted the files since I think we're not using them? If that's a mistake, please tell me...
Refactored the database opening and creation code so it's in a separate file.
