--- Mind-Full Meals/Mind-Full Meals/Calendar.storyboard
@@ -1,62 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<document type="com.apple.InterfaceBuilder3.CocoaTouch.Storyboard.XIB" version="3.0" toolsVersion="13771" targetRuntime="iOS.CocoaTouch" propertyAccessControl="none" useAutolayout="YES" useTraitCollections="YES" useSafeAreas="YES" colorMatched="YES">
-    <device id="retina4_7" orientation="portrait">
-        <adaptation id="fullscreen"/>
-    </device>
-    <dependencies>
-        <plugIn identifier="com.apple.InterfaceBuilder.IBCocoaTouchPlugin" version="13772"/>
-        <capability name="Safe area layout guides" minToolsVersion="9.0"/>
-        <capability name="documents saved in the Xcode 8 format" minToolsVersion="8.0"/>
-    </dependencies>
-    <scenes>
-        <!--Month-->
-        <scene sceneID="U3E-RO-OKQ">
-            <objects>
-                <viewController id="gmE-aj-irt" sceneMemberID="viewController">
-                    <view key="view" contentMode="scaleToFill" id="5H9-0m-xaJ">
-                        <rect key="frame" x="0.0" y="0.0" width="375" height="667"/>
-                        <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
-                        <color key="backgroundColor" white="1" alpha="1" colorSpace="calibratedWhite"/>
-                        <viewLayoutGuide key="safeArea" id="dNm-YC-OrT"/>
-                    </view>
-                    <navigationItem key="navigationItem" title="Month" id="O3n-i6-q0n"/>
-                </viewController>
-                <placeholder placeholderIdentifier="IBFirstResponder" id="7rp-Te-eJ1" userLabel="First Responder" sceneMemberID="firstResponder"/>
-            </objects>
-            <point key="canvasLocation" x="-209" y="-324"/>
-        </scene>
-        <!--View Controller-->
-        <scene sceneID="it0-iB-vip">
-            <objects>
-                <viewController id="PRw-mW-gs8" sceneMemberID="viewController">
-                    <view key="view" contentMode="scaleToFill" id="6rW-Bi-coQ">
-                        <rect key="frame" x="0.0" y="0.0" width="375" height="667"/>
-                        <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
-                        <color key="backgroundColor" white="1" alpha="1" colorSpace="calibratedWhite"/>
-                        <viewLayoutGuide key="safeArea" id="x4T-3U-oP9"/>
-                    </view>
-                </viewController>
-                <placeholder placeholderIdentifier="IBFirstResponder" id="ZoY-lK-jlU" userLabel="First Responder" sceneMemberID="firstResponder"/>
-            </objects>
-            <point key="canvasLocation" x="490" y="-324"/>
-        </scene>
-        <!--Navigation Controller-->
-        <scene sceneID="V4g-0A-S0j">
-            <objects>
-                <navigationController automaticallyAdjustsScrollViewInsets="NO" id="jKJ-aG-0HR" sceneMemberID="viewController">
-                    <toolbarItems/>
-                    <navigationBar key="navigationBar" contentMode="scaleToFill" insetsLayoutMarginsFromSafeArea="NO" id="IhJ-ea-3aJ">
-                        <rect key="frame" x="0.0" y="20" width="375" height="44"/>
-                        <autoresizingMask key="autoresizingMask"/>
-                    </navigationBar>
-                    <nil name="viewControllers"/>
-                    <connections>
-                        <segue destination="gmE-aj-irt" kind="relationship" relationship="rootViewController" id="Tmo-4U-NH7"/>
-                    </connections>
-                </navigationController>
-                <placeholder placeholderIdentifier="IBFirstResponder" id="9Ym-jg-qVw" userLabel="First Responder" sceneMemberID="firstResponder"/>
-            </objects>
-            <point key="canvasLocation" x="-646" y="-1049"/>
-        </scene>
-    </scenes>
-</document>
